receiver/requirements.txt
connexion[flask,uvicorn,swagger-ui]
pykafka
flask-cors
PyYAML
setuptools
